Move remixes page state to common
