+++ packages/web/src/common/store/pages/remixes/slice.ts
@@ -4,8 +4,8 @@ import { combineReducers } from 'redux'
 import { ID } from 'common/models/Identifiers'
 import { asLineup } from 'common/store/lineup/reducer'
 
-import { PREFIX as remixesTracksPrefix } from './lineups/tracks/actions'
-import remixesTracksReducer from './lineups/tracks/reducer'
+import { PREFIX as remixesTracksPrefix } from './lineup/actions'
+import remixesTracksReducer from './lineup/reducer'
 
 type State = {
   trackId: ID | null
