+++ packages/web/src/pages/remixes-page/store/lineups/tracks/sagas.ts
@@ -6,13 +6,13 @@ import { processAndCacheTracks } from 'common/store/cache/tracks/utils'
 import {
   PREFIX,
   tracksActions
-} from 'pages/remixes-page/store/lineups/tracks/actions'
-import { getTrackId, getLineup } from 'pages/remixes-page/store/selectors'
+} from 'common/store/pages/remixes/lineup/actions'
+import { getTrackId, getLineup } from 'common/store/pages/remixes/selectors'
 import apiClient from 'services/audius-api-client/AudiusAPIClient'
 import { LineupSagas } from 'store/lineup/sagas'
 import { AppState } from 'store/types'
 
-import { setCount } from '../../slice'
+import { setCount } from '../../../../../common/store/pages/remixes/slice'
 
 function* getTracks({
   offset,
