+++ packages/web/src/pages/remixes-page/store/sagas.ts
@@ -4,8 +4,12 @@ import { TrackMetadata } from 'common/models/Track'
 import { retrieveTrackByHandleAndSlug } from 'common/store/cache/tracks/utils/retrieveTracks'
 import { waitForBackendSetup } from 'store/backend/sagas'
 
+import {
+  fetchTrack,
+  fetchTrackSucceeded
+} from '../../../common/store/pages/remixes/slice'
+
 import tracksSagas from './lineups/tracks/sagas'
-import { fetchTrack, fetchTrackSucceeded } from './slice'
 
 function* watchFetch() {
   yield takeEvery(fetchTrack.type, function* (
