+++ packages/web/src/store/reducers.ts
@@ -20,7 +20,6 @@ import unfollowConfirmation from 'components/unfollow-confirmation-modal/store/r
 import dashboard from 'pages/artist-dashboard-page/store/reducer'
 import deactivateAccount from 'pages/deactivate-account-page/store/slice'
 import deleted from 'pages/deleted-page/store/slice'
-import remixes from 'pages/remixes-page/store/slice'
 import signOn from 'pages/sign-on/store/reducer'
 import upload from 'pages/upload-page/store/reducer'
 import visualizer from 'pages/visualizer/store/slice'
@@ -106,7 +105,6 @@ const createRootReducer = (routeHistory: History) =>
       pages: combineReducers({
         deleted,
         nowPlaying,
-        remixes,
         unfollowConfirmation
       })
     })
