+++ packages/web/src/store/types.ts
@@ -5,6 +5,7 @@ import averageColor from 'common/store/average-color/slice'
 import NotificationState from 'common/store/notifications/types'
 import { CollectionsPageState } from 'common/store/pages/collection/types'
 import HistoryPageState from 'common/store/pages/history-page/types'
+import RemixesPageReducer from 'common/store/pages/remixes/slice'
 import { SmartCollectionState } from 'common/store/pages/smart-collection/slice'
 import QueueReducer from 'common/store/queue/slice'
 import RemoteConfigReducer from 'common/store/remote-config/slice'
@@ -28,7 +29,6 @@ import { UnfollowConfirmationModalState } from 'components/unfollow-confirmation
 import ArtistDashboardState from 'pages/artist-dashboard-page/store/types'
 import { DeactivateAccountState } from 'pages/deactivate-account-page/store/slice'
 import DeletedPageReducer from 'pages/deleted-page/store/slice'
-import RemixesPageReducer from 'pages/remixes-page/store/slice'
 import SignOnPageState from 'pages/sign-on/store/types'
 import { UploadPageState } from 'pages/upload-page/store/types'
 import VisualizerReducer from 'pages/visualizer/store/slice'
